Plugins

+++ install.sh
@@ -1,24 +1,44 @@
 #!/bin/bash
 # Yanan Zhao, 2016
 
-# backup directory
-# with two leading dots to distinguish it from vim configuration files
-backup_dir=~/..vimback # with two leading dots to distinguish it from vim
+vim_runtime_dir=$HOME/.vim_runtime
+cur_vim_runtime_dir=$PWD/.vim_runtime
 
 # step.1. backup your old configurations
-echo "[1] backing up old configurations ..."
-if [ -d $backup_dir ]; then
-    echo "$backup_dir already exists, refresh it"
-    rm -rf $backup_dir
+echo -e "\033[36m[1] backing up old configurations ...\033[0m"
+if [ -d $vim_runtime_dir ]
+then
+    echo -e "$vim_runtime_dir exists, skiping backing up"
+else
+    # backup directory
+    # with two leading dots to distinguish it from vim configuration files
+    backup_dir=$HOME/..vimback # with two leading dots to distinguish it from vim
+
+    if [ -d $backup_dir ]; then
+        echo -e "$backup_dir already exists, refresh it"
+        rm -rf $backup_dir
+    fi
+
+    mkdir $backup_dir
+    mv $HOME/.vim* $backup_dir
+
+    echo -e "\033[34mbackup successful\033[0m"
 fi
+echo -e ""
 
-mkdir $backup_dir
-mv ~/.vim* $backup_dir
+# step.2. update plugins
+echo -e "\033[36m[2] updating plugins ...\033[0m"
+$cur_vim_runtime_dir/update_plugins.py
+if [ $? -gt 0 ]; then
+    echo -e "\033[31minstall abort\033[0m"
+    exit 1
+fi
+echo -e "\033[34mupdate successful\033[0m"
+echo -e ""
 
-echo "backup successful"
-echo ""
+# step.3. install
+echo -e "\033[36m[3] install new configurations ...\033[0m"
+find . -name ".ropeproject" -exec rm -rf {} +
+ln -sf $cur_vim_runtime_dir $vim_runtime_dir && sh $vim_runtime_dir/install_awesome_vimrc.sh
+echo -e "\033[34minstall finish\033[0m"
 
-# step.2. install
-echo "[2] install new configurations ..."
-cp -rf .vim_runtime ~ && sh ~/.vim_runtime/install_awesome_vimrc.sh
-echo "install finish"
